fix: Change ButtonType.Erro to Destructive to align with the Deisgn System
