+++ patterns/src/test/java/uk/gov/android/ui/patterns/errorscreen/v2/DestructiveScreenScrollingTest.kt
@@ -1,5 +1,7 @@
-package uk.gov.android.ui.patterns.errorscreen
+package uk.gov.android.ui.patterns.errorscreen.v2
 
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.assertIsNotDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
@@ -16,39 +18,51 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.RobolectricTestRunner
-import uk.gov.android.ui.patterns.centrealignedscreen.CentreAlignedScreenBodyContent
-import uk.gov.android.ui.patterns.centrealignedscreen.CentreAlignedScreenTestTag.BODY_LAZY_COLUMN_TEST_TAG
+import uk.gov.android.ui.componentsv2.heading.GdsHeading
+import uk.gov.android.ui.componentsv2.images.GdsIcon
+import uk.gov.android.ui.patterns.errorscreen.ErrorScreenIcon
+import uk.gov.android.ui.patterns.errorscreen.v2.ErrorScreenTitleTestTag.ERROR_BODY_LAZY_COLUMN_TEST_TAG
 import uk.gov.android.ui.patterns.utils.BDD.And
 import uk.gov.android.ui.patterns.utils.BDD.Given
 import uk.gov.android.ui.patterns.utils.BDD.Then
 import uk.gov.android.ui.patterns.utils.BDD.When
 import uk.gov.android.ui.patterns.utils.TestUtils.loremIpsum
+import uk.gov.android.ui.theme.spacingDouble
+import uk.gov.android.ui.theme.util.UnstableDesignSystemAPI
 
 @RunWith(RobolectricTestRunner::class)
-class ErrorScreenScrollingTest {
+class DestructiveScreenScrollingTest {
     @get:Rule
     val composeTestRule = createComposeRule()
 
     private lateinit var mandatoryIcon: ErrorScreenIcon
     private lateinit var mandatoryTitle: String
 
+    @OptIn(UnstableDesignSystemAPI::class)
     @Test
     fun `test scrolling behaviour in the body content`() = with(composeTestRule) {
         Given("a body content with multiple lines and paragraphs") {
             mandatoryIcon = ErrorScreenIcon.ErrorIcon
             mandatoryTitle = "Title"
         }
-        val topLine = CentreAlignedScreenBodyContent.Text("Top single line")
-        val paragraph = CentreAlignedScreenBodyContent.Text(loremIpsum(50))
-        val bottomLine = CentreAlignedScreenBodyContent.Text("Bottom single line")
+        val topLine = ErrorScreenBodyContent.Text("Top single line")
+        val paragraph = ErrorScreenBodyContent.Text(loremIpsum(50))
+        val bottomLine = ErrorScreenBodyContent.Text("Bottom single line")
         val body = persistentListOf(topLine, paragraph, paragraph, paragraph, paragraph, bottomLine)
 
         When("the screen is composed with the provided body") {
             setContent {
                 ErrorScreen(
-                    icon = mandatoryIcon,
-                    title = mandatoryTitle,
-                    body = body,
+                    icon = {
+                        GdsIcon(
+                            image = ImageVector.vectorResource(ErrorScreenIcon.ErrorIcon.icon),
+                            contentDescription = "Error",
+                        )
+                    },
+                    title = { GdsHeading(mandatoryTitle) },
+                    body = {
+                        toBodyContent(body, spacingDouble)
+                    },
                 )
             }
         }
@@ -62,7 +76,7 @@ class ErrorScreenScrollingTest {
         }
 
         When("body content is scrolled to the bottom") {
-            onNodeWithTag(BODY_LAZY_COLUMN_TEST_TAG)
+            onNodeWithTag(ERROR_BODY_LAZY_COLUMN_TEST_TAG)
                 .performTouchInput { swipeUp() }
                 .onChildren()
                 .onLast()
@@ -77,7 +91,7 @@ class ErrorScreenScrollingTest {
         }
 
         When("body content is scrolled back to the top") {
-            onNodeWithTag(BODY_LAZY_COLUMN_TEST_TAG)
+            onNodeWithTag(ERROR_BODY_LAZY_COLUMN_TEST_TAG)
                 .performTouchInput { swipeDown() }
                 .onChildren()
                 .onFirst()
